add changes based on 1.x
