+++ src/Asset/Controller/Image/CustomDownloadController.php
@@ -14,18 +14,20 @@ declare(strict_types=1);
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\StudioBackendBundle\Asset\Controller\Download\Image;
+namespace Pimcore\Bundle\StudioBackendBundle\Asset\Controller\Image;
 
 use OpenApi\Attributes\Get;
 use Pimcore\Bundle\StudioBackendBundle\Asset\Attributes\Response\Content\AssetMediaType;
 use Pimcore\Bundle\StudioBackendBundle\Asset\Attributes\Response\Header\ContentDisposition;
-use Pimcore\Bundle\StudioBackendBundle\Asset\OpenApi\Attributes\Parameters\Path\FormatTypeParameter;
+use Pimcore\Bundle\StudioBackendBundle\Asset\MappedParameter\ImageDownloadConfigParameter;
+use Pimcore\Bundle\StudioBackendBundle\Asset\OpenApi\Attributes\Parameters\Query\ImageConfigParameter;
+use Pimcore\Bundle\StudioBackendBundle\Asset\OpenApi\Attributes\Parameters\Query\MimeTypeParameter;
+use Pimcore\Bundle\StudioBackendBundle\Asset\OpenApi\Attributes\Parameters\Query\ResizeModeParameter;
 use Pimcore\Bundle\StudioBackendBundle\Asset\Service\AssetServiceInterface;
 use Pimcore\Bundle\StudioBackendBundle\Asset\Service\DownloadServiceInterface;
 use Pimcore\Bundle\StudioBackendBundle\Controller\AbstractApiController;
 use Pimcore\Bundle\StudioBackendBundle\Exception\AccessDeniedException;
 use Pimcore\Bundle\StudioBackendBundle\Exception\ElementNotFoundException;
-use Pimcore\Bundle\StudioBackendBundle\Exception\InvalidAssetFormatTypeException;
 use Pimcore\Bundle\StudioBackendBundle\Exception\InvalidElementTypeException;
 use Pimcore\Bundle\StudioBackendBundle\Exception\ThumbnailResizingFailedException;
 use Pimcore\Bundle\StudioBackendBundle\OpenApi\Attributes\Parameters\Path\IdParameter;
@@ -35,13 +37,14 @@ use Pimcore\Bundle\StudioBackendBundle\OpenApi\Config\Tags;
 use Pimcore\Bundle\StudioBackendBundle\Security\Service\SecurityServiceInterface;
 use Pimcore\Bundle\StudioBackendBundle\Util\Constants\HttpResponseCodes;
 use Symfony\Component\HttpFoundation\BinaryFileResponse;
+use Symfony\Component\HttpKernel\Attribute\MapQueryString;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\Serializer\SerializerInterface;
 
 /**
  * @internal
  */
-final class FormatDownloadController extends AbstractApiController
+final class CustomDownloadController extends AbstractApiController
 {
     public function __construct(
         private readonly AssetServiceInterface $assetService,
@@ -55,43 +58,50 @@ final class FormatDownloadController extends AbstractApiController
     /**
      * @throws AccessDeniedException
      * @throws ElementNotFoundException
-     * @throws InvalidAssetFormatTypeException
      * @throws InvalidElementTypeException
      * @throws ThumbnailResizingFailedException
      */
     #[Route(
-        '/assets/{id}/image/download/{format}',
-        name: 'pimcore_studio_api_download_image_format',
+        '/assets/{id}/image/download/custom',
+        name: 'pimcore_studio_api_download_image_custom',
         methods: ['GET']
     )]
     //#[IsGranted('STUDIO_API')]
     //#[IsGranted(UserPermissions::ASSETS->value)]
     #[Get(
-        path: self::API_PATH . '/assets/{id}/image/download/{format}',
-        operationId: 'downloadImageByFormat',
-        description: 'Download image by id and format by path parameters',
-        summary: 'Download image by id and format',
+        path: self::API_PATH . '/assets/{id}/image/download/custom',
+        operationId: 'downloadCustomImage',
+        description: 'Download custom image by configuration',
+        summary: 'Download custom image by configuration',
         security: self::SECURITY_SCHEME,
         tags: [Tags::Assets->name]
     )]
-    #[IdParameter(type: 'asset')]
-    #[FormatTypeParameter]
+    #[IdParameter(type: 'image')]
+    #[MimeTypeParameter]
+    #[ResizeModeParameter]
+    #[ImageConfigParameter('width', 140)]
+    #[ImageConfigParameter('height')]
+    #[ImageConfigParameter('quality')]
+    #[ImageConfigParameter('dpi')]
     #[SuccessResponse(
-        description: 'Image based on format',
-        content: new AssetMediaType('image/jpeg'),
+        description: 'Custom image',
+        content: [new AssetMediaType('image/jpeg'), new AssetMediaType('image/png')],
         headers: [new ContentDisposition()]
     )]
     #[DefaultResponses([
         HttpResponseCodes::UNAUTHORIZED,
         HttpResponseCodes::NOT_FOUND,
     ])]
-    public function downloadImageByFormat(int $id, string $format): BinaryFileResponse
+    public function downloadCustomImage(
+        int $id,
+        #[MapQueryString] ImageDownloadConfigParameter $parameters
+    ): BinaryFileResponse
     {
         $asset = $this->assetService->getAssetElement(
             $this->securityService->getCurrentUser(),
             $id
         );
 
-        return $this->downloadService->downloadImageByFormat($asset, $format);
+        return $this->downloadService->downloadCustomImage($asset, $parameters);
     }
 }
