+++ src/Asset/Service/DownloadService.php
@@ -22,6 +22,8 @@ use Pimcore\Bundle\StudioBackendBundle\Exception\InvalidAssetFormatTypeException
 use Pimcore\Bundle\StudioBackendBundle\Exception\InvalidElementTypeException;
 use Pimcore\Bundle\StudioBackendBundle\Exception\ThumbnailResizingFailedException;
 use Pimcore\Bundle\StudioBackendBundle\Util\Constants\Asset\FormatTypes;
+use Pimcore\Bundle\StudioBackendBundle\Util\Constants\HttpResponseHeaders;
+use Pimcore\Bundle\StudioBackendBundle\Util\Traits\StreamedResponseTrait;
 use Pimcore\Model\Asset;
 use Pimcore\Model\Asset\Image;
 use Pimcore\Model\Element\ElementInterface;
@@ -33,6 +35,8 @@ use Symfony\Component\HttpFoundation\StreamedResponse;
  */
 final readonly class DownloadService implements DownloadServiceInterface
 {
+    use StreamedResponseTrait;
+
     public function __construct(
         private ThumbnailServiceInterface $thumbnailService,
         private array $defaultFormats,
@@ -51,21 +55,7 @@ final readonly class DownloadService implements DownloadServiceInterface
             throw new InvalidElementTypeException($asset->getType());
         }
 
-        $stream = $asset->getStream();
-        if (!is_resource($stream)) {
-            throw new ElementStreamResourceNotFoundException(
-                $asset->getId(),
-                'Asset'
-            );
-        }
-
-        return new StreamedResponse(function () use ($stream) {
-            fpassthru($stream);
-        }, 200, [
-            'Content-Type' => $asset->getMimeType(),
-            'Content-Disposition' => sprintf('attachment; filename="%s"', $asset->getFilename()),
-            'Content-Length' => $asset->getFileSize(),
-        ]);
+        return $this->getStreamedResponse($asset, HttpResponseHeaders::ATTACHMENT_TYPE->value);
     }
 
     /**
@@ -103,12 +93,12 @@ final readonly class DownloadService implements DownloadServiceInterface
             throw new InvalidAssetFormatTypeException($format);
         }
         $parameters = new ImageDownloadConfigParameter(
-            $configuration['mimeType'],
-            $configuration['resizeMode'],
-            $configuration['width'],
-            $configuration['height'],
-            $configuration['quality'],
-            $configuration['dpi']
+            mimeType: $configuration['format'],
+            resizeMode: $configuration['resize_mode'],
+            width: $configuration['width'] ?? null,
+            height: $configuration['height'] ?? null,
+            quality: $configuration['quality'] ?? null,
+            dpi: $configuration['dpi'] ?? null
         );
 
         return $this->thumbnailService->getBinaryResponseFromThumbnail(
