+++ src/Util/Constants/Thumbnails.php
@@ -1,0 +1,26 @@
+<?php
+declare(strict_types=1);
+
+/**
+ * Pimcore
+ *
+ * This source file is available under two different licenses:
+ * - GNU General Public License version 3 (GPLv3)
+ * - Pimcore Commercial License (PCL)
+ * Full copyright and license information is available in
+ * LICENSE.md which is distributed with this source code.
+ *
+ *  @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
+ *  @license    http://www.pimcore.org/license     GPLv3 and PCL
+ */
+
+namespace Pimcore\Bundle\StudioBackendBundle\Util\Constants;
+
+/**
+ * @internal
+ */
+enum Thumbnails: string
+{
+    case DEFAULT_THUMBNAIL_ID = 'pimcore-system-treepreview';
+    case DEFAULT_THUMBNAIL_TEXT = 'original';
+}
